requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
static3==0.7.0
urllib3==1.26.11
whitenoise==6.2.0
